refactor(Table): [Country] Refactor country table type and pagination filters sorting type, update loading on table and search style
