+++ src/components/ui/Inputs/Search/searchInput.module.scss
@@ -1,6 +1,9 @@
 .search_input {
   width: 20%;
   position: relative;
+  &:focus-visible {
+    outline-color: var(--primary) !important;
+  }
   .icon {
     position: absolute;
     top: 8px;
