+++ src/types/Country.ts
@@ -1,0 +1,12 @@
+export interface Continent {
+  id: number;
+  name: string;
+}
+
+export interface Country {
+  id: number;
+  iso: string;
+  nicename: string;
+  isParticipate: boolean;
+  continent: Continent;
+}
