+++ src/types/Filters.ts
@@ -4,3 +4,9 @@ export interface BasicFilters {
   sorts: string | null;
   [key: string]: string | number | null; // To allow additional filter properties
 }
+
+export interface CountryFilters extends BasicFilters {
+  name: string | null;
+  continentId: string | null;
+  isParticipate: string | null;
+}
