+++ src/types/Sort.ts
@@ -1,0 +1,4 @@
+export interface Sort {
+  key: string;
+  order: "asc" | "desc";
+}
